refactor: 패키지 구조 변경
- 도메인과 서비스 로직 나눔
