+++ src/main/java/pairmatching/model/MatchMachine.java
@@ -2,6 +2,9 @@ package pairmatching.model;
 
 import java.util.List;
 
+import pairmatching.model.domain.History;
+import pairmatching.model.domain.Options;
+
 public class MatchMachine {
     private final MatchGenerator generator;
     private final List<History> historys;
