+++ src/main/java/pairmatching/model/domain/History.java
@@ -1,4 +1,4 @@
-package pairmatching.model;
+package pairmatching.model.domain;
 
 import java.util.List;
