+++ src/main/java/pairmatching/model/domain/Level.java
@@ -1,6 +1,5 @@
-package pairmatching.model;
+package pairmatching.model.domain;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
