+++ src/main/java/pairmatching/model/domain/Pair.java
@@ -1,8 +1,9 @@
-package pairmatching.model;
+package pairmatching.model.domain;
 
 import java.util.ArrayList;
 import java.util.List;
 
+import pairmatching.model.domain.Crew;
 import pairmatching.util.Constant;
 
 public class Pair {
